+++ apps/decodex/src/orchestrator/tests/operator/status/history/attention/ledger.rs
@@ -1,0 +1,55 @@
+use crate::orchestrator::tests::operator::status::{
+	self, FakeTracker, StateStore, TEST_SERVICE_ID, orchestrator, slice,
+};
+
+#[test]
+fn live_operator_history_lanes_require_linear_execution_ledger_records() {
+	let (_temp_dir, config, workflow) = status::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = status::sample_issue_with_sort_fields(
+		"issue-1",
+		"XY-356",
+		"Done",
+		&[],
+		Some(3),
+		"2026-04-29T10:11:00Z",
+	);
+	let tracker = FakeTracker::new(vec![issue.clone()]);
+
+	state_store
+		.upsert_worktree(
+			TEST_SERVICE_ID,
+			&issue.id,
+			"y/decodex-xy-356",
+			&config.worktree_root().join(&issue.identifier).display().to_string(),
+		)
+		.expect("worktree should remember project ownership");
+	state_store
+		.record_run_attempt("xy-356-attempt-1", &issue.id, 1, "succeeded")
+		.expect("successful attempt should record");
+	state_store
+		.clear_worktree(&issue.id)
+		.expect("completed lane cleanup should clear local worktree");
+
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("snapshot should build");
+	let lane = snapshot.history_lanes.first().expect("history lane should exist");
+	let rendered = orchestrator::render_operator_status(&snapshot);
+
+	assert_eq!(tracker.comment_queries.borrow().as_slice(), slice::from_ref(&issue.id));
+	assert_eq!(lane.ledger_outcome.ledger_status, "missing");
+	assert_eq!(lane.ledger_outcome.final_outcome, "execution_ledger_missing");
+	assert_eq!(lane.ledger_outcome.record_count, 0);
+	assert_eq!(
+		lane.ledger_outcome.summary.as_deref(),
+		Some("No decodex.linear_execution_event records are available for this history lane.")
+	);
+	assert!(rendered.contains("ledger_status: missing"));
+	assert!(rendered.contains("outcome: execution_ledger_missing"));
+}
